requirements_server.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
importlib-resources==5.12.0
kiwisolver==1.4.4
numpy==1.24.2
packaging==23.0
paho-mqtt==1.6.1
Pillow==9.4.0
psutil==5.9.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-uinput==0.11.2
python-xlib==0.33
six==1.16.0
zipp==3.15.0